Uploading files under 100MBs
Because GitHUb can only allow 100MB commits and pushes at a time i started through to get the most number of files into the the repo at this starting time. The big exclusions that took the most space were the modes and art folders and the metafiles folder in the library folder.