Updated Boss and Level 1 Scenes(Nav Mesh, Collision, Character movement, and more)
With this committ this sets up alot of what is needed for ai and other key components. The player character works and collides in both scenes.  Return of the old terrain with updated nav mesh and working collision. scaling for boss scene and for player character. Finally, Off mesh links set in the boss scene and just need ai to use. (PS. This all took me 10 hours straight with no breaks to get implemented with all the trial and error.)